+++ CHANGELOG.md
@@ -3,3 +3,4 @@ All Notable changes to `ElementFinder` will be documented in this file
 
 ## Version 0.0.3
 - Feature #4 Use `DOMAttr::nodeValue` instead of `DOMAttr::value`    
+- BC #7 Refactor `Helper` class. Create `FormHelper`, `NodeHelper` and `StringHelper`